api/requirements.txt
celery==5.0.5
Django==3.2.1
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
numpy==1.20.2
pandas==1.2.4
pytest-cov==2.12.1
pytest-django==4.3.0
mixer==7.1.2
pytest-mock==3.6.1
redis==3.5.3
pylint==2.8.3
pylint-django==2.4.4
